[server] Mock image upload/retrieval routes
Mocked POST /image/upload/:bucket route which uploads an image to a specified bucket and returns its ID and GET /image/:bucket/:id which gets an image from a specified bucket with its ID.

To parse the file data in the request, added multer as a dependency to act as middleware. The mocked data currently supports getting images from the "profile" and "organization" buckets, using random image generator services.